feat: add custom hook to fetch data
